github_project_uploader/requirements.txt
# requirements.txt
Flask
Flask-Dance
python-dotenv
requests
gunicorn
GitPython